+++ app/services/bender-bot/main.go
@@ -4,14 +4,14 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/zmoog/go-bender/business/bot"
+	"github.com/zmoog/go-bender/business/bot/commands"
 	"github.com/zmoog/go-bender/foundation/logger"
-
-	"github.com/zmoog/go-bender/bot"
-	"github.com/zmoog/go-bender/bot/commands"
-	"github.com/zmoog/go-bender/scraper/jsonscraper"
+	"github.com/zmoog/go-bender/foundation/scraper/jsonscraper"
 )
 
 func main() {
+	// Logging
 	log, err := logger.New("bender")
 	if err != nil {
 		fmt.Printf("Error creating logger: %v\n", err)
@@ -19,6 +19,7 @@ func main() {
 	}
 	defer log.Sync()
 
+	// Configuration
 	token, ok := os.LookupEnv("DISCORD_TOKEN")
 	if !ok {
 		log.Error("DISCORD_TOKEN environment variable not set")
@@ -27,15 +28,19 @@ func main() {
 
 	log.Info("Starting bot")
 
+	// Dependencies
 	jsonscraper := jsonscraper.New()
 
+	// Commands
 	bender := bot.New(log, token)
 	bender.AddCommand(commands.ListAppleProducts(jsonscraper))
 
+	// Startup
 	err = bender.Run()
 	if err != nil {
 		log.Errorw("Error running bot: %v", err)
 	}
 
+	// Shutdown
 	log.Info("Bot stopped")
 }
